// backend/interfaces/APIError.ts
export enum APIError {
  UNAUTHORIZED = "UNAUTHORIZED",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  TOO_EARLY = "TOO_EARLY",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  INVALID_REQUEST = "INVALID_REQUEST",
  UNSUPPORTED_URL_SCHEME = "UNSUPPORTED_URL_SCHEME",
  RESOURCE_EXCEEDS_FILE_SIZE = "RESOURCE_EXCEEDS_FILE_SIZE",
  FILE_NOT_FOUND = "FILE_NOT_FOUND",
}